backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain-anthropic
anthropic
google-genai
langfuse
python-dotenv
pydantic
supabase
pytest
